aula_3/T3/flask_user_api/requirements.txt
flask==3.0.0
flask-restful==0.3.10
validate-docbr==1.10.0
python-dateutil==2.8.2